blogger_web/app/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Login
Flask-Migrate
Flask-WTF
Flask-Caching
prometheus-flask-exporter
psycopg2-binary
gunicorn
redis
email_validator
pillow
